feat: ahhachul.com - filter management with api call
