+++ services/ahhachul.com/src/pages/communicate/data/index.ts
@@ -1,0 +1,5 @@
+export const COMMUNITY_FILTER_DEFAULT_VALUES = {
+  category: 'HOT',
+  line: 'allLines',
+  search: '',
+};
