+++ services/ahhachul.com/src/pages/communicate/ui/Page/Page.tsx
@@ -1,8 +1,9 @@
 import React, { useReducer } from 'react';
 import { useActivity } from '@stackflow/react';
 import type { ActivityComponentType } from 'app/stackflow';
+import { COMMUNITY_FILTER_DEFAULT_VALUES } from 'pages/communicate/data';
 import { Layout } from 'widgets';
-import { ComposeLayout } from 'widgets/layout/ui/ComposeLayout';
+import { FilterProvider } from 'widgets/filters/ui/FilterContext';
 import { renderLeftLogo, renderRight } from 'widgets/layout-header';
 import { ArticleListErrorFallback } from 'widgets/articles/ui/ArticleListErrorFallback';
 import { QueryErrorBoundary } from 'entities/app-errors/ui/QueryErrorBoundary';
@@ -18,7 +19,10 @@ const Community: ActivityComponentType = () => {
   const [isScale, toggleScale] = useReducer((scale) => !scale, false);
 
   return (
-    <ComposeLayout>
+    <FilterProvider
+      id="community"
+      defaultValues={COMMUNITY_FILTER_DEFAULT_VALUES}
+    >
       <CommunityFilters
         isScale={isScale}
         isActive={activity.isActive}
@@ -43,7 +47,7 @@ const Community: ActivityComponentType = () => {
           <CommunityArticleList css={styles.layout(isScale)} />
         </QueryErrorBoundary>
       </Layout>
-    </ComposeLayout>
+    </FilterProvider>
   );
 };
