+++ services/ahhachul.com/src/pages/communicate/ui/_common/CommunityArticleList/CommunityArticleList.tsx
@@ -5,17 +5,18 @@ import type { CommunityArticle } from 'pages/communicate/model';
 import { Loading } from 'entities/app-loaders/ui/Loading';
 import { flattenInfinityList } from 'shared/lib/utils/array/flattenInfinityList';
 import { BaseArticleList } from 'widgets/articles/ui/BaseArticleList';
-import { useCommunityFilterStore } from 'pages/communicate/slice';
+import { useFilters } from 'widgets/filters/ui/FilterContext';
 
 interface CommunityArticleListProps
   extends HTMLAttributes<HTMLSectionElement> {}
 const CommunityArticleList = ({ ...props }: CommunityArticleListProps) => {
-  const { filters } = useCommunityFilterStore((state) => state);
+  const { keyword, filters } = useFilters();
   const { data, hasNextPage, isFetchingNextPage, fetchNextPage } =
     useGetCommunityList({
       page: 0,
       size: 10,
       sort: 'createdAt,asc',
+      content: keyword,
       ...filters,
     });
   const communityArticles = flattenInfinityList(data);
@@ -33,5 +34,5 @@ const CommunityArticleList = ({ ...props }: CommunityArticleListProps) => {
 };
 
 export default withSuspense(CommunityArticleList, {
-  fallback: <Loading opacity={1} />,
+  fallback: <Loading opacity={1} deferredMs={0} />,
 });
