+++ services/ahhachul.com/src/pages/lost-found/api/get-list.ts
@@ -23,4 +23,6 @@ export const useGetLostFoundList = (params: ParamsOfLostFoundList) =>
       getLostFoundList({ ...params, page: pageParam }),
     getNextPageParam: (lastPage) => lastPage.data.result.nextPageNum,
     initialPageParam: params?.page ?? 0,
+    gcTime: 5 * 60 * 1000,
+    staleTime: 30 * 1000,
   });
