+++ services/ahhachul.com/src/pages/lost-found/data/index.ts
@@ -1,0 +1,5 @@
+export const LOST_FOUND_FILTER_DEFAULT_VALUES = {
+  type: 'LOST',
+  line: 'allLines',
+  search: '',
+};
