+++ services/ahhachul.com/src/pages/lost-found/ui/Page/Page.tsx
@@ -2,12 +2,13 @@ import React, { useReducer } from 'react';
 import { useActivity } from '@stackflow/react';
 import type { ActivityComponentType } from '@stackflow/react';
 import { Layout } from 'widgets';
-import { ComposeLayout } from 'widgets/layout/ui/ComposeLayout';
 import { renderLeftLogo, renderRight } from 'widgets/layout-header';
+import { FilterProvider } from 'widgets/filters/ui/FilterContext';
 import { ArticleListErrorFallback } from 'widgets/articles/ui/ArticleListErrorFallback';
 import { QueryErrorBoundary } from 'entities/app-errors/ui/QueryErrorBoundary';
 import { LostFoundFilters } from '../_common/LostFoundFilters/LostFoundFilters';
 import * as styles from './Page.css';
+import { LOST_FOUND_FILTER_DEFAULT_VALUES } from 'pages/lost-found/data';
 
 const LostFoundArticleList = React.lazy(
   () => import('../_common/LostFoundArticleList/LostFoundArticleList'),
@@ -18,7 +19,10 @@ const LostFound: ActivityComponentType = () => {
   const [isScale, toggleScale] = useReducer((scale) => !scale, false);
 
   return (
-    <ComposeLayout>
+    <FilterProvider
+      id="lostFound"
+      defaultValues={LOST_FOUND_FILTER_DEFAULT_VALUES}
+    >
       <LostFoundFilters
         isScale={isScale}
         isActive={activity.isActive}
@@ -37,7 +41,7 @@ const LostFound: ActivityComponentType = () => {
           <LostFoundArticleList css={styles.layout(isScale)} />
         </QueryErrorBoundary>
       </Layout>
-    </ComposeLayout>
+    </FilterProvider>
   );
 };
